requirements.txt
absl-py==0.9.0
alembic==1.4.2
appdirs==1.4.3
astroid==2.4.1
astunparse==1.6.3
Babel==2.8.0
bcrypt==3.1.7
bert-for-tf2==0.14.4
blinker==1.4
CacheControl==0.12.6
cachetools==4.1.0
certifi==2019.11.28
cffi==1.14.0
chardet==3.0.4
click==7.1.2
colorama==0.4.3
contextlib2==0.6.0
cryptography==2.9.2
distlib==0.3.0
distro==1.4.0
filelock==3.0.12
flake8==3.8.1
Flask==1.0.2
Flask-BabelEx==0.9.4
Flask-Compress==1.4.0
Flask-Gravatar==0.5.0
Flask-JWT-Extended==3.24.1
Flask-Login==0.4.1
Flask-Mail==0.9.1
Flask-Migrate==2.4.0
Flask-Paranoid==0.2.0
Flask-Principal==0.4.0
Flask-Script==2.0.6
Flask-Security==3.0.0
Flask-SQLAlchemy==2.3.2
Flask-WTF==0.14.2
gast==0.3.3
gdown==3.11.0
google-api-core==1.17.0
google-auth==1.15.0
google-auth-httplib2==0.0.3
google-auth-oauthlib==0.4.1
google-pasta==0.2.0
googleapis-common-protos==1.51.0
grpcio==1.30.0
h5py==2.10.0
html5lib==1.0.1
httplib2==0.18.1
idna==2.8
importlib-metadata==1.6.0
ipaddr==2.2.0
isort==4.3.21
itsdangerous==1.1.0
Jinja2==2.11.2
joblib==0.14.1
Keras==2.3.1
Keras-Applications==1.0.8
Keras-Preprocessing==1.1.2
lazy-object-proxy==1.4.3
lockfile==0.12.2
Mako==1.1.2
Markdown==3.2.2
MarkupSafe==1.1.1
mccabe==0.6.1
msgpack==0.6.2
nltk==3.5
numpy==1.18.4
oauthlib==3.1.0
opt-einsum==3.2.1
packaging==20.3
paramiko==2.7.1
params-flow==0.8.2
passlib==1.7.1
pep517==0.8.2
pgadmin4==4.17
progress==1.5
protobuf==3.12.2
psutil==5.5.1
psycopg2==2.8.5
psycopg2-binary==2.8.5
punctuator==0.9.5
py-params==0.9.7
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycodestyle==2.6.0
pycparser==2.20
pyflakes==2.2.0
PyJWT==1.7.1
pylint==2.5.2
PyNaCl==1.3.0
pyparsing==2.4.6
PySocks==1.7.1
python-dateutil==2.8.1
python-editor==1.0.4
pytoml==0.1.21
pytz==2018.9
PyYAML==5.3.1
regex==2020.5.7
requests==2.22.0
requests-oauthlib==1.3.0
retrying==1.3.3
rsa==4.0
scipy==1.4.1
sentencepiece==0.1.91
simplejson==3.16.0
six==1.14.0
speaklater==1.3
SQLAlchemy==1.3.16
sqlparse==0.2.4
sshtunnel==0.1.5
tensorboard==2.2.2
tensorboard-plugin-wit==1.7.0
tensorflow==2.2.0
tensorflow-estimator==2.2.0
tensorflow-gpu==2.2.0
termcolor==1.1.0
Theano==1.0.4
toml==0.10.0
tqdm==4.46.0
typed-ast==1.4.1
uritemplate==3.0.1
urllib3==1.25.8
uWSGI==2.0.18
webencodings==0.5.1
Werkzeug==1.0.1
wrapt==1.12.1
WTForms==2.2.1
zipp==3.1.0
